feat(layout): move streak chip to top header
Surface the daily streak prominently in the top-right header as a pill
chip (flame icon + count) so it's visible on every screen. Sidebar
becomes nav-only for a cleaner shell. Mobile MomentumStrip on the
practice page still shows the full streak/attempts/due breakdown.
